Add resis. conv test to diffusion one
